backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
httpx==0.27.0
beautifulsoup4==4.12.3
apscheduler==3.10.4
python-dotenv==1.0.1
lxml==5.3.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
